work-in-progress snapshot of GSet refactor
