+++ source/vstd/vstd.rs
@@ -59,6 +59,8 @@ pub mod relations;
 pub mod rwlock;
 pub mod seq;
 pub mod seq_lib;
+pub mod gset;
+pub mod iset;
 pub mod set;
 pub mod set_lib;
 pub mod shared;
